+++ src/WinGetUtilInterop.UnitTests/ManifestUnitTest/V1ManifestReadTest.cs
@@ -197,6 +197,21 @@ namespace WinGetUtilInterop.UnitTests.ManifestUnitTest
             Assert.Equal(2, manifest.InstallerSuccessCodes.Count);
             Assert.Equal(1, manifest.InstallerSuccessCodes[0]);
             Assert.Equal(0x80070005, manifest.InstallerSuccessCodes[1]);
+
+            if (manifestVersion >= TestManifestVersion.V1_12_0)
+            {
+                var defaultUninstallerSwitches = manifest.UninstallerSwitches;
+                Assert.Equal("/custom", defaultUninstallerSwitches.Custom);
+                Assert.Equal("/silentwithprogress", defaultUninstallerSwitches.SilentWithProgress);
+                Assert.Equal("/silence", defaultUninstallerSwitches.Silent);
+                Assert.Equal("/interactive", defaultUninstallerSwitches.Interactive);
+                Assert.Equal("/log=<LOGPATH>", defaultUninstallerSwitches.Log);
+
+                Assert.Equal(2, manifest.UninstallerSuccessCodes.Count);
+                Assert.Equal(2, manifest.UninstallerSuccessCodes[0]);
+                Assert.Equal(0x80070002, manifest.UninstallerSuccessCodes[1]);
+            }
+
             Assert.Equal("uninstallPrevious", manifest.UpgradeBehavior);
             Assert.Equal(2, manifest.Commands.Count);
             Assert.Equal("makemsix", manifest.Commands[0]);
@@ -423,6 +438,16 @@ namespace WinGetUtilInterop.UnitTests.ManifestUnitTest
                 Assert.Equal("microsoftEntraId", installer1.Authentication.AuthenticationType);
                 Assert.Equal("Resource", installer1.Authentication.MicrosoftEntraIdAuthenticationInfo.Resource);
                 Assert.Equal("Scope", installer1.Authentication.MicrosoftEntraIdAuthenticationInfo.Scope);
+            }
+
+            if (manifestVersion >= TestManifestVersion.V1_12_0)
+            {
+                var installer1UninstallerSwitches = installer1.UninstallerSwitches;
+                Assert.Equal("/c", installer1UninstallerSwitches.Custom);
+                Assert.Equal("/sp", installer1UninstallerSwitches.SilentWithProgress);
+                Assert.Equal("/s", installer1UninstallerSwitches.Silent);
+                Assert.Equal("/i", installer1UninstallerSwitches.Interactive);
+                Assert.Equal("/l=<LOGPATH>", installer1UninstallerSwitches.Log);
             }
 
             // Additional Localizations